+++ tensorlayer/decorators/deprecated.py
@@ -1,6 +1,7 @@
 #! /usr/bin/python
 # -*- coding: utf-8 -*-
 
+import inspect
 import sys
 import functools
 
@@ -8,8 +9,6 @@ from tensorflow.python.util import decorator_utils
 
 from tensorflow.python.util.deprecation import _call_location
 from tensorflow.python.util.deprecation import _validate_deprecation_args
-
-from tensorlayer import logging
 
 import wrapt
 
@@ -83,10 +82,13 @@ def deprecated(wrapped=None, date='', instructions='', warn_once=True):
                 if warn_once:
                     _PRINTED_WARNING[class_or_func_name] = True
 
+                from tensorlayer import logging
+
                 logging.warning(
-                    'From %s: %s (from %s) is deprecated and will be removed %s.\n'
+                    '%s: `%s.%s` (in file: %s) is deprecated and will be removed %s.\n'
                     'Instructions for updating: %s\n' % (
-                        _call_location(), class_or_func_name, wrapped.__module__, 'in a future version'
+                        "Class" if inspect.isclass(wrapped) else "Function", wrapped.__module__, class_or_func_name,
+                        wrapped.__code__.co_filename, 'in a future version'
                         if date is None else ('after %s' % date), instructions
                     )
                 )